requirements.txt
ccxt
pandas
numpy
APScheduler
SQLAlchemy
python-dotenv
rich